Backend/requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0
django-allauth==0.39.1
django-default-imagefield==0.5.5
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.1
gcloud==0.18.3
googleapis-common-protos==1.5.10
gunicorn==19.9.0
httplib2==0.12.3
idna==2.8
jws==0.1.3
oauth2client==4.1.2
oauthlib==3.0.1
Pillow==5.4.1
protobuf==3.8.0rc1
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodome==3.8.1
Pyrebase4==4.3.0
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
stripe==2.27.0
urllib3==1.24.3
whitenoise==4.1.2